FILEUPLOAD-302 Remove junit.framework.TestCase dependency
junit.framework.TestCase was introduced in JUnit 3, and while it
isn't technically deprecated, it's outdated, and should not be used
in JUnit 4.

This patch removes its usage in the project, and replaces it with
static imports from org.junit.Assert, as per JUnit 4's best
practices.

In addition, `@Test` annotations were added to the test methods, so
they are recognized as tests by JUnit 4.